+++ application/modules/editor/Models/Import/FileParser/Xlf/SurroundingTagRemover/Paired.php
@@ -30,85 +30,17 @@ END LICENSE AND COPYRIGHT
  * calculates and removes leading and trailing paired and special single tags
  * removing means: the tags are not imported, they are added directly to the skeleton file and are not changeable therefore
  */
-class editor_Models_Import_FileParser_Xlf_SurroundingTags {
-    
-    /**
-     * @var Zend_Config
-     */
-    protected $config;
-    
-    /**
-     * @var boolean
-     */
-    protected $preserveWhitespace;
-    
-    private $startShiftCount = 0;
-    private $endShiftCount = 0;
-    
-    protected $leadingTags = '';
-    protected $trailingTags = '';
+class editor_Models_Import_FileParser_Xlf_SurroundingTagRemover_Paired extends editor_Models_Import_FileParser_Xlf_SurroundingTagRemover_Abstract {
     
     /**
      * the following single tags are also to be considered to be moved out of the segment,
      * since they are place holders for isolated paired tags
      * @var array
      */
-    protected $isolatedPairedTags = ['it', 'bx', 'ex'];
-    
-    public function __construct(Zend_Config $config) {
-        $this->config = $config;
-    }
-    
-    /**
-     * calculates the tags to be cut off
-     * @param boolean $preserveWhitespace
-     * @param array $sourceChunks
-     * @param array $targetChunks
-     * @param editor_Models_Import_FileParser_XmlParser $xmlparser
-     */
-    public function calculate(bool $preserveWhitespace, array $sourceChunks, array $targetChunks, editor_Models_Import_FileParser_XmlParser $xmlparser) {
-        $this->leadingTags = '';
-        $this->trailingTags = '';
-        $this->preserveWhitespace = $preserveWhitespace;
-        
-        //if target is empty, we assume the target = source so that the feature can be used at all
-        // 1. because it checks for same tags in source and target
-        // 2. because we need the tags from source to be added as leading / trailing in target
-        $target = $xmlparser->join($targetChunks);
-        if(empty($target) && $target !== "0") {
-            $targetChunks = $sourceChunks;
-        }
-        
-        //reset start/end shift count.
-        // the counts are set by hasSameStartAndEndTags to > 0,
-        // then the start/end offset where the placeHolder is placed is shifted
-        // to exclude tags leading and trailing tags in the segment
-        $this->startShiftCount = 0;
-        $this->endShiftCount = 0;
-        //if preserveWhitespace is enabled, hasSameStartAndEndTags should not hide tags,
-        // since potential whitespace tags does matter then in the content
-        //since we are calling the leading/trailing tag stuff on the already fetched source segments,
-        // we have no ability here to conserve content outside the mrk tags - which also should not be on preserveWhitespace
-        if(!$this->hasSameStartAndEndTags($sourceChunks, $targetChunks)) {
-            //if there is just leading/trailing whitespace but no tags we reset the counter
-            // since then we dont want to cut off something
-            //if there is whitespace between or before the leading / after the trailing tags,
-            // this whitespace is ignored depending the preserveWhitespace setting.
-            // above $sourceChunks $targetChunks does not contain any irrelevant whitespace (only empty chunks)
-            $this->startShiftCount = 0;
-            $this->endShiftCount = 0;
-            return;
-        }
-        
-        $prevMode = editor_Models_Import_FileParser_Tag::setMode(editor_Models_Import_FileParser_Tag::RETURN_MODE_ORIGINAL);
-        //we get and store the leading target tags for later insertion
-        $this->leadingTags = $xmlparser->join(array_slice($targetChunks, 0, $this->startShiftCount));
-        
-        //we get and store the trailing target tags for later insertion
-        if($this->endShiftCount > 0) {
-            $this->trailingTags = $xmlparser->join(array_slice($targetChunks, -$this->endShiftCount));
-        }
-        editor_Models_Import_FileParser_Tag::setMode($prevMode);
+    protected array $isolatedPairedTags = ['it', 'bx', 'ex'];
+
+    protected function _calculate(array $sourceChunks, array $targetChunks): bool {
+        return $this->hasSameStartAndEndTags($sourceChunks, $targetChunks);
     }
     
     /**
@@ -120,10 +52,9 @@ class editor_Models_Import_FileParser_Xlf_SurroundingTags {
      * @param bool $foundTag used for recursive call
      * @return boolean returns false if there are no matching leading/trailing tags at all
      */
-    protected function hasSameStartAndEndTags(array $source, array $target, $foundTag = false) {
+    protected function hasSameStartAndEndTags(array $source, array $target, bool $foundTag = false): bool {
         //source and target must have at least a start or end tag and inbetween text content, that means at least 2 chunks:
-        // if the feature is disabled no framing tags are ignored
-        if(!$this->config->runtimeOptions->import->xlf->ignoreFramingTags || count($source) < 3 || count($target) < 3){
+        if(count($source) < 3 || count($target) < 3){
             return $foundTag;
         }
         
@@ -201,10 +132,10 @@ class editor_Models_Import_FileParser_Xlf_SurroundingTags {
      * @param bool $fromStart true if trim from start, false to trim from end
      * @return int  returns -1 if source and targetChunk do not match!
      */
-    protected function trimAndCount(&$sourceChunk, array $source, array $target, $fromStart = true): int {
+    protected function trimAndCount(mixed &$sourceChunk, array $source, array $target, bool $fromStart = true): int {
         $shiftCount = 0;
-        
-        while(!is_null($sourceChunk) && empty($sourceChunk) || (!$this->preserveWhitespace && preg_match('#^\s+$#', $sourceChunk))) {
+
+        while(!is_null($sourceChunk) && strlen($sourceChunk) == 0 || (!$this->preserveWhitespace && preg_match('#^\s+$#', $sourceChunk))) {
             if($fromStart) {
                 $sourceChunk = array_shift($source);
                 $targetChunk = array_shift($target);
@@ -220,32 +151,7 @@ class editor_Models_Import_FileParser_Xlf_SurroundingTags {
             //inc internal start shift count
             $shiftCount++;
         }
-        
+
         return $shiftCount;
-    }
-    
-    /**
-     * removes the leading and trailing tags as calculated before
-     * @param array $chunks
-     * @return array
-     */
-    public function sliceTags(array $chunks): array {
-        return array_slice($chunks, $this->startShiftCount, count($chunks) - $this->startShiftCount - $this->endShiftCount);
-    }
-    
-    /**
-     * get the leading cut of tags in original import format as string
-     * @return string
-     */
-    public function getLeading(): string {
-        return $this->leadingTags;
-    }
-    
-    /**
-     * get the trailing cut of tags in original import format as string
-     * @return string
-     */
-    public function getTrailing(): string {
-        return $this->trailingTags;
     }
 }