+++ application/modules/editor/testcases/editorAPI/Translate2282Test.php
@@ -51,7 +51,7 @@ class Translate2282Test extends editor_Test_JsonTest {
         
         $tests = array(
             'runtimeOptions.import.xlf.preserveWhitespace' => 0,
-            'runtimeOptions.import.xlf.ignoreFramingTags' => 1,
+            'runtimeOptions.import.xlf.ignoreFramingTags' => 'all',
         );
         self::$api->testConfig($tests);